userview alerts

+++ frontend/src/Pages/Profile/shopcard/ShopCard.jsx
@@ -1,4 +1,3 @@
-
 import React, { useState, useEffect } from "react";
 import "./shopcard.css";
 import CIcon from "@coreui/icons-react";
@@ -28,7 +27,6 @@ const ShopCard = ({ shoppost, onDelete }) => {
       try {
         const userData = await fetchUserData(shoppost.postedBy);
         setAuthor(userData);
-        console.log("The author is", userData);
       } catch (error) {
         console.error("Error fetching author:", error);
       }
@@ -43,60 +41,53 @@ const ShopCard = ({ shoppost, onDelete }) => {
     return null;
   }
 
-
   const handleDelete = () => {
     setShowDeleteAlert(true);
   };
+  
 
   const handleAlertClose = () => {
+    if (onDelete && typeof onDelete === 'function') {
+      onDelete(shoppost._id);
+    }
     setShowDeleteAlert(false);
   };
 
-  const confirmDelete = () => {
-    onDelete(shoppost._id);
-    setShowDeleteAlert(false);
-  };
-
+ 
   return (
-    
-    <div className="shoppingcard"> 
-    <Link
-            key={shoppost._id} // Key should be placed on the outermost JSX element inside map
-            style={{ textDecoration: "none" }}
-            to={`/productdescription/${shoppost._id}`}
-          >
-      <img src={shoppost.imageUrl} alt="" /></Link>
+    <div className="shoppingcard">
+      <Link
+        key={shoppost._id} // Key should be placed on the outermost JSX element inside map
+        style={{ textDecoration: "none" }}
+        to={`/productdescription/${shoppost._id}`}
+      >
+        <img src={shoppost.imageUrl} alt="" />
+      </Link>
       <div className="shopcardtext">
         <div className="shopcardtitle">{shoppost.name}</div>
         <div className="shopcardprice">{shoppost.price}</div>
-         <div className="shopprofile">
-        <button onClick={handleDelete} className="shop-card-delete-button" aria-label="Delete post">
-          <CIcon
-            icon={icon.cilTrash}
-            size=""
-            style={{ "--ci-primary-color": "black" }}
-          />
-        </button>
-          {author&&(
+        <div className="shopprofile">
+          <button onClick={handleDelete} className="shop-card-delete-button" aria-label="Delete post">
+            <CIcon
+              icon={icon.cilTrash}
+              size=""
+              style={{ "--ci-primary-color": "black" }}
+            />
+          </button>
+          {author && (
             <div className="shopprofilepic">
-              <img src={author.profilePicture} alt={author.username} className="authorProfilePicture"/>
+              <img src={author.profilePicture} alt={author.username} className="authorProfilePicture" />
             </div>
-
           )}
-          
         </div>
-        
       </div>
       {showDeleteAlert && (
         <Alert
           message="Are you sure you want to delete this post?"
           onClose={handleAlertClose}
-          onConfirm={confirmDelete}
+          
         />
       )}
-   
-        
-      
     </div>
   );
 };
